% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisguru-package.R
\docType{package}
\name{apisguru-package}
\alias{apisguru}
\alias{apisguru-package}
\title{apisguru: APIs.guru API Client}
\description{
A client to interact with the APIs.guru API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/apisguru}
  \item \url{https://jonthegeek.github.io/apisguru/}
  \item Report bugs at \url{https://github.com/jonthegeek/apisguru/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
