% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_shared.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters used in multiple functions}
\arguments{
\item{api}{The version of the API.}

\item{provider}{The provider of the API, such as "1password.com" or
"apis.guru".}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
