% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-shared.R
\name{.shared-params}
\alias{.shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{req}{(\code{httr2_request}) The request object to modify.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
