% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-apis-get_api.R
\name{get_api}
\alias{get_api}
\alias{req_get_api}
\alias{tidy_policy_get_api}
\title{Retrieve one version of a particular API}
\usage{
get_api(provider, version, max_reqs = Inf, max_tries_per_req = 3)

req_get_api(provider, version)

tidy_policy_get_api()
}
\arguments{
\item{provider}{(length-1 \code{character})}

\item{version}{(length-1 \code{character})}

\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}
}
\value{
\code{get_api()}: The API response.

\code{req_get_api()}: (\code{httr2_request}) A \code{\link[httr2:request]{httr2::request()}} object.

\code{tidy_policy_get_api()}: (\code{nectar_tidy_policy}) A list like the ones
returned by \code{\link[nectar:tidy_policy_prepare]{nectar::tidy_policy_prepare()}}.
}
\description{
Returns the API entry for one specific version of an API.
}
