% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-apis-get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\alias{req_get_metrics}
\alias{tidy_policy_get_metrics}
\title{Get basic metrics}
\usage{
get_metrics(max_reqs = Inf, max_tries_per_req = 3)

req_get_metrics()

tidy_policy_get_metrics()
}
\arguments{
\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}
}
\value{
\code{get_metrics()}: The API response.

\code{req_get_metrics()}: (\code{httr2_request}) A \code{\link[httr2:request]{httr2::request()}} object.

\code{tidy_policy_get_metrics()}: (\code{nectar_tidy_policy}) A list like the
ones returned by \code{\link[nectar:tidy_policy_prepare]{nectar::tidy_policy_prepare()}}.
}
\description{
Some basic metrics for the entire \href{https://apis.guru}{apis.guru} directory.
}
