% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-apis-get_provider.R
\name{get_provider}
\alias{get_provider}
\alias{req_get_provider}
\alias{tidy_policy_get_provider}
\title{List all APIs for a particular provider}
\usage{
get_provider(provider, max_reqs = Inf, max_tries_per_req = 3)

req_get_provider(provider)

tidy_policy_get_provider()
}
\arguments{
\item{provider}{(length-1 \code{character})}

\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}
}
\value{
\code{get_provider()}: Information about the API entry for each API
provided by the \code{provider}.

\code{req_get_provider()}: (\code{httr2_request}) A \code{\link[httr2:request]{httr2::request()}} object.

\code{tidy_policy_get_provider()}: (\code{nectar_tidy_policy}) A list like the
ones returned by \code{\link[nectar:tidy_policy_prepare]{nectar::tidy_policy_prepare()}}.
}
\description{
List all APIs in the \href{https://apis.guru}{apis.guru} directory for a
particular provider. Returns links to the individual API entry for each API.
}
