% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-apis-get_providers.R
\name{get_providers}
\alias{get_providers}
\alias{req_get_providers}
\alias{tidy_policy_get_providers}
\title{List all providers}
\usage{
get_providers(max_reqs = Inf, max_tries_per_req = 3)

req_get_providers()

tidy_policy_get_providers()
}
\arguments{
\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}
}
\value{
\code{get_providers()}: The API response.

\code{req_get_providers()}: (\code{httr2_request}) A \code{\link[httr2:request]{httr2::request()}}
object.

\code{tidy_policy_get_providers()}: (\code{nectar_tidy_policy}) A list like
the ones returned by \code{\link[nectar:tidy_policy_prepare]{nectar::tidy_policy_prepare()}}.
}
\description{
List all the providers in the \href{https://apis.guru}{apis.guru} directory.
}
