% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-apis-get_service_api.R
\name{get_service_api}
\alias{get_service_api}
\alias{req_get_service_api}
\alias{tidy_policy_get_service_api}
\title{Retrieve one version of a particular API with a serviceName.}
\usage{
get_service_api(
  provider,
  service,
  version,
  max_reqs = Inf,
  max_tries_per_req = 3
)

req_get_service_api(provider, service, version)

tidy_policy_get_service_api()
}
\arguments{
\item{provider}{(length-1 \code{character})}

\item{service}{(length-1 \code{character})}

\item{version}{(length-1 \code{character})}

\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}
}
\value{
\code{get_service_api()}: The API response.

\code{req_get_service_api()}: (\code{httr2_request}) A \code{\link[httr2:request]{httr2::request()}}
object.

\code{tidy_policy_get_service_api()}: (\code{nectar_tidy_policy}) A list like
the ones returned by \code{\link[nectar:tidy_policy_prepare]{nectar::tidy_policy_prepare()}}.
}
\description{
Returns the API entry for one specific version of an API where there is a
serviceName.
}
