% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-call.R
\name{guru_call_api}
\alias{guru_call_api}
\title{Call the APIs.guru API}
\usage{
guru_call_api(path, query = NULL, body = NULL, method = NULL)
}
\arguments{
\item{path}{The route to an API endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{An optional list of parameters to pass in the query portion of
the request.}

\item{body}{An object to use as the body of the request. If any component of
the body is a path, pass it through \code{\link[fs:path]{fs::path()}} or otherwise give it the
class "fs_path" to indicate that it is a path.}

\item{method}{If the method is something other than GET or POST, supply it.
Case is ignored.}
}
\value{
The response from the endpoint.
}
\description{
Generate a request to an APIs.guru endpoint.
}
