% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_api.R
\name{guru_get_api}
\alias{guru_get_api}
\title{Retrieve one version of a particular API}
\usage{
guru_get_api(parms = stop("Get these from the spec"))
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
Returns the API entry for one specific version of an API where there is no
serviceName.
}
