% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_metrics.R
\name{guru_get_metrics}
\alias{guru_get_metrics}
\title{Get basic metrics}
\usage{
guru_get_metrics()
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
Some basic metrics for the entire directory. Just stunning numbers to put on
a front page and are intended purely for WoW effect :)
}
