% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_provider.R
\name{guru_get_provider}
\alias{guru_get_provider}
\title{List all APIs for a particular provider}
\usage{
guru_get_provider(parms = stop("Get these from the spec"))
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
List all APIs in the directory for a particular providerName Returns links to
the individual API entry for each API.
}
