% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_services.R
\name{guru_get_services}
\alias{guru_get_services}
\title{List all serviceNames for a particular provider}
\usage{
guru_get_services(parms = stop("Get these from the spec"))
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
List all serviceNames in the directory for a particular providerName
}
