% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-prepare.R
\name{guru_req_prepare}
\alias{guru_req_prepare}
\title{Generate a request for the APIs.guru API}
\usage{
guru_req_prepare(
  path,
  query = list(),
  body = NULL,
  header = list(),
  cookie = list(),
  method = NULL,
  tidy_policy = nectar::tidy_policy_unknown(),
  call = rlang::caller_env()
)
}
\arguments{
\item{path}{(\code{character} or \code{list}) The route to an API endpoint. Optionally,
a list or character vector with the path as one or more unnamed arguments
(which will be concatenated with "/") plus named arguments to
\code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{(\code{character} or \code{list}) An optional list or character vector of
parameters to pass in the query portion of the request. Can also include a
\code{.multi} argument to pass to \code{\link[httr2:req_url_query]{httr2::req_url_query()}} to control how
elements containing multiple values are handled.}

\item{body}{(multiple types) An object to use as the body of the request. If
any component of the body is a path, pass it through \code{\link[fs:path]{fs::path()}} or
otherwise give it the class "fs_path" to indicate that it is a path.}

\item{header}{(\code{list} or \code{NULL}) An optional list of headers to add to the
request using \code{\link[httr2:req_headers]{httr2::req_headers()}}. A \code{NULL} value for an individual
header will explicitly remove that header if it was previously set.}

\item{cookie}{(\code{list} or \code{NULL}) An optional list of cookies to set on the
request using \code{\link[httr2:req_cookies_set]{httr2::req_cookies_set()}}. \code{NULL} elements are removed.}

\item{method}{(\verb{length-1 character}, optional) If the method is something
other than \code{GET} or \code{POST}, supply it. Case is ignored.}

\item{tidy_policy}{(\code{nectar_tidy_policy} or \code{NULL}) A tidying policy prepared
with \code{\link[nectar:tidy_policy_prepare]{tidy_policy_prepare()}}. By default, \code{\link[nectar:tidy_policy_body_auto]{tidy_policy_body_auto()}} is used
to automatically apply \code{\link[nectar:resp_body_auto]{resp_body_auto()}} to responses.}

\item{call}{(\code{environment}) The environment from which a function was called,
e.g. \code{\link[rlang:caller_env]{rlang::caller_env()}} (the default). The environment will be mentioned
in error messages as the source of the error. This argument is particularly
useful for functions that are intended to be called as utilities inside
other functions.}
}
\value{
A \code{\link[httr2:request]{httr2::request()}} object with additional class \code{nectar_request}.
}
\description{
Prepare a request for the APIs.guru API, using the opinionated framework
defined in \code{\link[nectar:req_init]{nectar::req_init()}}, \code{\link[nectar:req_modify]{nectar::req_modify()}},
\code{\link[nectar:req_tidy_policy]{nectar::req_tidy_policy()}}, and \code{\link[nectar:req_pagination_policy]{nectar::req_pagination_policy()}}.
}
\details{
You may wish to export this function (if the API changes often or you do not
fully implement the API, for example).
}
\keyword{internal}
