% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-apis-list_apis.R
\name{list_apis}
\alias{list_apis}
\alias{req_list_apis}
\alias{tidy_policy_list_apis}
\title{List all APIs}
\usage{
list_apis(max_reqs = Inf, max_tries_per_req = 3)

req_list_apis()

tidy_policy_list_apis()
}
\arguments{
\item{max_reqs}{(length-1 \code{integer}) The maximum number of separate requests
to perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(length-1 \code{integer}) The maximum number of times to
attempt each individual request. Passed on to
\code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}
}
\value{
\code{list_apis()}: The API response.

\code{req_list_apis()}: (\code{httr2_request}) A \code{\link[httr2:request]{httr2::request()}} object.

\code{tidy_policy_list_apis()}: (\code{nectar_tidy_policy}) A list like the
ones returned by \code{\link[nectar:tidy_policy_prepare]{nectar::tidy_policy_prepare()}}.
}
\description{
List all APIs in the \href{https://apis.guru}{apis.guru} directory. Returns links
to the OpenAPI definitions for each API in the directory. If API exist in
multiple versions \code{preferred} one is explicitly marked. Some basic info from
the OpenAPI definition is cached inside each object. This allows you to
generate some simple views without needing to fetch the OpenAPI definition
for each API.
}
