% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_api.R
\name{schema_api}
\alias{schema_api}
\title{Meta information about API}
\usage{
schema_api(x)
}
\arguments{
\item{x}{A list with elements \code{added}, \code{preferred}, and \code{versions}.}
}
\value{
A one-row tibble describing the API.
}
\description{
Apply the API schema to the details of a single API.
}
\examples{
schema_api()
x <- list(
  added = "2016-01-01T12:34:56Z",
  preferred = "v2",
  versions = list(
    v1 = list(
      added = "2016-01-01T12:34:56Z",
      info = list(),
      updated = "2016-01-01T12:34:56Z",
      swaggerUrl = "https://example.com/v1/swagger.json",
      swaggerYamlUrl = "https://example.com/v1/swagger.yml",
      openapiVer = "3.0.0"
    ),
    v2 = list(
      added = "2023-01-01T12:34:56Z",
      info = list(),
      updated = "2023-01-01T12:34:56Z",
      swaggerUrl = "https://example.com/v2/swagger.json",
      swaggerYamlUrl = "https://example.com/v2/swagger.yml",
      openapiVer = "3.1.0"
    )
  )
)
schema_api(x)
}
