% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_apis.R
\name{schema_apis}
\alias{schema_apis}
\title{Meta information about APIs}
\usage{
schema_apis(x = list())
}
\arguments{
\item{x}{A named list of API details. Each element should have elements
\code{added}, \code{preferred}, and \code{versions}.}
}
\value{
A tibble with one row per API.
}
\description{
Apply the APIs schema to a list of API details.
}
\examples{
schema_apis()
x <- list(
  "api1" = list(
    added = "2016-01-01T12:34:56Z",
    preferred = "v2",
    versions = list(
      v1 = list(
        added = "2016-01-01T12:34:56Z",
        info = list(),
        updated = "2016-01-01T12:34:56Z",
        swaggerUrl = "https://example.com/v1/swagger.json",
        swaggerYamlUrl = "https://example.com/v1/swagger.yml",
        openapiVer = "3.0.0"
      ),
      v2 = list(
        added = "2023-01-01T12:34:56Z",
        info = list(),
        updated = "2023-01-01T12:34:56Z",
        swaggerUrl = "https://example.com/v2/swagger.json",
        swaggerYamlUrl = "https://example.com/v2/swagger.yml",
        openapiVer = "3.1.0"
      )
    )
  ),
  "api2" = list(
    added = "2018-01-01T12:34:56Z",
    preferred = "v1",
    versions = list(
      v1 = list(
        added = "2018-01-01T12:34:56Z",
        info = list(),
        updated = "2018-01-01T12:34:56Z",
        swaggerUrl = "https://url.com/v1/swagger.json",
        swaggerYamlUrl = "https://url.com/v1/swagger.yml",
        openapiVer = "3.0.0"
      )
    )
  )
)
schema_apis(x)
}
