% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_metrics.R
\name{schema_metrics}
\alias{schema_metrics}
\title{List of basic metrics}
\usage{
schema_metrics(x)
}
\arguments{
\item{x}{A list of metrics, with elements \code{numSpecs}, \code{numAPIs}, and
\code{numEndpoints} (and likely several other elements).}
}
\value{
A validated list of metrics.
}
\description{
Apply the Metrics schema a metrics object.
}
