% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-call.R
\name{call_guru_api}
\alias{call_guru_api}
\title{Call the APIs.guru API}
\usage{
call_guru_api(path)
}
\arguments{
\item{path}{The route to an API endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}
}
\value{
The response from the endpoint.
}
\description{
Generate a request to an APIs.guru endpoint.
}
