% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_shared.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters used in multiple functions}
\arguments{
\item{api}{A string giving the version of the API, such as "2.2.0" or "v2".}

\item{provider}{A string describing the API provider, such as such as
"1password.com" or "apis.guru".}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
