% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_api.R
\name{get_api}
\alias{get_api}
\title{Retrieve one version of a particular API}
\usage{
get_api(provider, api)
}
\arguments{
\item{provider}{A string describing the API provider, such as such as
"1password.com" or "apis.guru".}

\item{api}{A string giving the version of the API, such as "2.2.0" or "v2".}
}
\value{
A \code{\link[=schema_api_version]{schema_api_version()}} tibble.
}
\description{
Returns the API entry for one specific version of an API where there is no
serviceName.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_api("apis.guru", "2.2.0")
\dontshow{\}) # examplesIf}
}
