% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_api.R
\name{get_api}
\alias{get_api}
\title{Retrieve one version of a particular API}
\usage{
get_api(provider, api)
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
Returns the API entry for one specific version of an API where there is no
serviceName.
}
