% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get basic metrics}
\usage{
get_metrics()
}
\value{
A \code{\link[=schema_metrics]{schema_metrics()}} list object.
}
\description{
Some basic metrics for the entire directory. Just stunning numbers to put on
a front page and are intended purely for WoW effect :)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_metrics()
\dontshow{\}) # examplesIf}
}
