% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_provider.R
\name{get_provider}
\alias{get_provider}
\title{List all APIs for a particular provider}
\usage{
get_provider(provider)
}
\arguments{
\item{provider}{The provider of the API, such as "1password.com" or
"apis.guru".}
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
List all APIs in the directory for a particular providerName. Returns links to
the individual API entry for each API.
}
