% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_provider.R
\name{get_provider}
\alias{get_provider}
\title{List all APIs for a particular provider}
\usage{
get_provider(provider)
}
\arguments{
\item{provider}{A string describing the API provider, such as such as
"1password.com" or "apis.guru".}
}
\value{
A \code{\link[=schema_api_versions]{schema_api_versions()}} tibble.
}
\description{
List all APIs in the directory for a particular providerName. Returns links
to the individual API entry for each API.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
head(get_provider())
\dontshow{\}) # examplesIf}
}
