% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_providers.R
\name{get_providers}
\alias{get_providers}
\title{List all providers}
\usage{
get_providers()
}
\value{
A character vector of provider names.
}
\description{
List all the providers in the directory
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
head(get_providers())
\dontshow{\}) # examplesIf}
}
