% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_service_api.R
\name{get_service_api}
\alias{get_service_api}
\title{Retrieve one version of a particular API with a serviceName.}
\usage{
get_service_api(parms = stop("Get these from the spec"))
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
Returns the API entry for one specific version of an API where there is a serviceName.
}
