% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_services.R
\name{get_services}
\alias{get_services}
\title{List all serviceNames for a particular provider}
\usage{
get_services(provider)
}
\arguments{
\item{provider}{A string describing the API provider, such as such as
"1password.com" or "apis.guru".}
}
\value{
A character vector of service names.
}
\description{
List all serviceNames in the directory for a particular providerName
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_services("1password.com")
get_services("adyen.com")
\dontshow{\}) # examplesIf}
}
