% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_get_services.R
\name{get_services}
\alias{get_services}
\title{List all serviceNames for a particular provider}
\usage{
get_services(provider)
}
\arguments{
\item{provider}{The provider of the API, such as "1password.com" or
"apis.guru".}
}
\value{
FILL THIS FROM COMPONENTS
}
\description{
List all serviceNames in the directory for a particular providerName
}
