% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_list_apis.R
\name{list_apis}
\alias{list_apis}
\title{List all APIs}
\usage{
list_apis()
}
\value{
A \code{\link[=schema_apis]{schema_apis()}} tibble.
}
\description{
List all APIs in the directory. Returns links to the OpenAPI definitions for
each API in the directory. If API exist in multiple versions \code{preferred} one
is explicitly marked. Some basic info from the OpenAPI definition is cached
inside each object. This allows you to generate some simple views without
needing to fetch the OpenAPI definition for each API.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
list_apis()
\dontshow{\}) # examplesIf}
}
