% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_api_version.R
\name{schema_api_version}
\alias{schema_api_version}
\title{Description of one version of an API}
\usage{
schema_api_version(x)
}
\arguments{
\item{x}{An object describing one version of an API.}
}
\value{
A one-row tibble describing the API.
}
\description{
Apply the ApiVersions schema to an API object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- call_guru_api("/specs/apis.guru/2.2.0.json")
names(x)
schema_api_version(x)
\dontshow{\}) # examplesIf}
}
